feat : pod template 生成方法重构
