+++ api/v1/consts.go
@@ -1,0 +1,6 @@
+package v1
+
+const (
+	DefaultTopologyKey string = `kubernetes.io/hostname`
+	FlinkNativeType    string = `flink-native-kubernetes`
+)
