+++ api/v1/flinksession_types_generate_test.go
@@ -2,7 +2,37 @@ package v1
 
 import (
 	"github.com/stretchr/testify/assert"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"testing"
+)
+
+const (
+	Expected = `apiVersion: v1
+kind: Pod
+metadata:
+  creationTimestamp: null
+spec:
+  affinity:
+    podAntiAffinity:
+      preferredDuringSchedulingIgnoredDuringExecution:
+      - podAffinityTerm:
+          labelSelector:
+            matchExpressions:
+            - key: app
+              operator: In
+              values:
+              - flink
+            - key: type
+              operator: In
+              values:
+              - flink-native-kubernetes
+          namespaces:
+          - test
+          topologyKey: kubernetes.io/hostname
+        weight: 100
+  containers: []
+status: {}
+`
 )
 
 func TestBuildNodeSelector(t *testing.T) {
@@ -25,4 +55,19 @@ func TestBuildNodeSelector(t *testing.T) {
 	if res != "flink:run,disk:ssd" && res != "disk:ssd,flink:run" {
 		assert.Fail(t, "map 拼接失败！")
 	}
+}
+
+func TestPodTemplate(t *testing.T) {
+	session := &FlinkSession{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "flink",
+			Namespace: "test",
+		},
+		Spec: FlinkSessionSpec{
+			BalancedSchedule: PreferredDuringScheduling,
+		},
+	}
+
+	template := session.GeneratePodTemplate()
+	assert.Equal(t, Expected, template)
 }
